Optimised Game of life using numpy
We now store our grid on a numpy array. We then pad this array with 0s and shift it in the 8 compass directions. Then we sum up the shifted grids to calculate the number of neighbours for each cell. Lastly we use a custom numpy ufunc to update our array and display.